In select/cons_res fix bug that could result in a job being allocated zero     CPUs on some nodes. Patch from Hongjia Cao, NUDT.

+++ trunk/src/plugins/select/cons_res/job_test.c
@@ -1183,7 +1183,7 @@ static int _eval_nodes(struct job_record *job_ptr, bitstr_t *node_map,
 				if ((details_ptr->shared != 0) &&
 				    (avail_cpus > rem_cpus)) {
 					avail_cpus = MAX(rem_cpus,
-							 details_ptr->
+							 (int)details_ptr->
 							 pn_min_cpus);
 					cpu_cnt[i] = MAX(avail_cpus, 1);
 				}
@@ -1221,7 +1221,7 @@ static int _eval_nodes(struct job_record *job_ptr, bitstr_t *node_map,
 				if ((details_ptr->shared != 0) &&
 				    (avail_cpus > rem_cpus)) {
 					avail_cpus = MAX(rem_cpus,
-							 details_ptr->
+							 (int)details_ptr->
 							 pn_min_cpus);
 					cpu_cnt[i] = MAX(avail_cpus, 1);
 				}
@@ -1305,7 +1305,7 @@ static int _eval_nodes(struct job_record *job_ptr, bitstr_t *node_map,
 				if ((details_ptr->shared != 0) &&
 				    (avail_cpus > rem_cpus)) {
 					avail_cpus = MAX(rem_cpus,
-							 details_ptr->
+							 (int)details_ptr->
 							 pn_min_cpus);
 					cpu_cnt[i] = MAX(avail_cpus, 1);
 				}
